typechecker: refactor into multiple files
